pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-harness"
version = "0.1.0"
description = "AI Agent 개발 워크플로 하네스 — 기획부터 배포까지 8단계 오케스트레이터"
readme = "README.md"
requires-python = ">=3.10"
keywords = ["ai", "agent", "harness", "workflow", "claude", "codex"]
dependencies = [
    "typer[all]>=0.12.0",
    "rich>=13.7.0",
    "pydantic>=2.6.0",
    "pyyaml>=6.0.1",
    "requests>=2.31.0",
    "python-dotenv>=1.0.1",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "jinja2>=3.1.4",
    "python-multipart>=0.0.9",
    "markdown>=3.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.27.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "bandit>=1.7.7",
]

[project.scripts]
harness = "harness.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["harness*"]

# ── ruff (lint + format) ──────────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "S"]
ignore = ["S101", "S603", "S607"]   # assert, subprocess 허용

# ── mypy ──────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true

# ── pytest ────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts   = "-v --tb=short"
